refactor: Enhance delete logic to prevent deletion of categories linked to dishes or meal sets
- Updated category deletion logic to check for associations with dishes or meal sets
- Added a business exception that is thrown if an attempt is made to delete a category that is linked to active dishes or meal sets

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -2,12 +2,16 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import org.springframework.beans.BeanUtils;
@@ -26,6 +30,10 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private DishMapper dishMapper;
+    @Autowired
+    private SetmealMapper setmealMapper;
 
     /**
      * 新增分类
@@ -102,6 +110,15 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public void remove(Long id) {
+        // 查询当前分类是否关联了菜品，如果关联了就抛出业务异常，不能删除
+        if (dishMapper.countByCategoryId(id) > 0)
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+
+        // 查询当前分类是否关联了套餐，如果关联了就抛出业务异常，不能删除
+        if (setmealMapper.countByCategoryId(id) > 0)
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+
+        // 删除分类
         categoryMapper.delete(id);
     }
